Resolve "Add eCMR data model v1.0.0 to the backend and adjust backend due to data model changes"

+++ src/main/java/org/openlogisticsfoundation/ecmr/domain/beans/ItemBean.java
@@ -20,6 +20,6 @@ class ItemBean {
     private Integer logisticsPackageItemQuantity;
     private String logisticsPackageType;
     private String transportCargoIdentification;
-    private Integer supplyChainConsignmentItemGrossWeight;
-    private Integer supplyChainConsignmentItemGrossVolume;
+    private Float supplyChainConsignmentItemGrossWeight;
+    private Float supplyChainConsignmentItemGrossVolume;
 }
